# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: API de gestion des articles
  description: Documentation de l'API pour la gestion des articles.
  version: 1.0.0
servers:
  - url: /api
    description: Serveur de développement local
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
paths:
...
  /v1/users:
    get:
      summary: Obtenir la liste des utilisateurs
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des utilisateurs récupérée avec succès
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
...
  /v1/users-by-etat:
    post:
      summary: Obtenir les utilisateurs par état
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des utilisateurs par état
  /v1/clients/filter-by-telephone:
    post:
      summary: Filtrer les clients par téléphone
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Liste des clients filtrés par téléphone
...